feat: Add iteration count to handleAddIteration function
This code change adds the iteration count to the handleAddIteration function in the AIAgentSidebar component. The iteration count is calculated based on the length of the IterationsList and is used to set the number and name properties of the newIteration object. This enhancement improves the accuracy and consistency of iteration numbering in the application.